调整 raft cluster model结构
