+++ src/raft/cluster/mod.rs
@@ -1,2 +1,3 @@
 pub mod route;
 pub mod routeapi;
+pub mod model;
